重构物品输出逻辑，支持最高单种long数量的输出

+++ src/main/java/com/EyeOfHarmonyBuffer/ItemInfo.java
@@ -4,11 +4,11 @@ public class ItemInfo {
 
     public String modid; // 模组 ID
     public String itemName; // 物品的注册名
-    public int quantity; // 输出数量
+    public long quantity; // 输出数量
     public int meta; // 物品的元数据值
     public String oreDictName; // 矿物词典
 
-    public ItemInfo(String modid, String itemName, int quantity, int meta) {
+    public ItemInfo(String modid, String itemName,long quantity, int meta) {
         this.modid = modid;
         this.itemName = itemName;
         this.quantity = quantity;
@@ -16,7 +16,7 @@ public class ItemInfo {
         this.oreDictName = null;
     }
 
-    public ItemInfo(String oreDictName, Integer quantity) {
+    public ItemInfo(String oreDictName, long quantity) {
         this.modid = null;
         this.itemName = null;
         this.quantity = quantity;
